% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{convert_list_col}
\alias{convert_list_col}
\title{Convert a dataframe within a dataframe to columns}
\usage{
convert_list_col(data_to_concat, data, name_col)
}
\arguments{
\item{name_col}{}
}
\description{
Will work with different lengths of of columns
}
