% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_base_temp}
\alias{crop_base_temp}
\title{Crop Base Temperature Dataset}
\format{
a dataframe of 20 rows and 5 columns
\describe{
\item{crop_name}{Name of Crop}
\item{base_temp_f}{Lower threshold of crop in Farenheit}
\item{base_temp_c}{Lower threshold of crop in Celsius}
\item{upper_threshold_f}{Upper threshold of crop in Farenheit}
\item{upper_threshold_c}{Upper threshold of crop in Celsius}
}
}
\source{
a csv file give to use
}
\usage{
crop_base_temp
}
\description{
A dataset containing the lower and upper threshold in Fahrenheit and Celsius for 20 crops Stotz's is interested
}
\keyword{datasets}
