% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fields_250}
\alias{fields_250}
\title{Fields 250}
\format{
a dataframe of 250 rows and 5 columns
\describe{
\item{FIELD_ID}{Field ID}
\item{field_lat}{Latitude of the field}
\item{field_lon}{Longitude of the field}
\item{state}{State the field is in}
\item{field_elevation}{Elevation of the field measured in feet}
}
}
\usage{
fields_250
}
\description{
A dataset containing 250 Field ID's of interest for Stotz and their associated Latitude, Longitude, State, and field elevation
}
\keyword{datasets}
