% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_calculations.R
\name{gdh}
\alias{gdh}
\title{Calculates the Growing Degree Hours}
\usage{
gdh(..., crop, crop_data)
}
\arguments{
\item{crop}{The name of the crop}

\item{crop_data}{The specific growing degree range. So the Minimum and Maximum degrees}
}
\value{
A numeric value
}
\description{
Gets the summation of the growing degree hours for a specific crop.
}
