% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{get_chunk_info}
\alias{get_chunk_info}
\title{Getting specified field's metadata}
\usage{
get_chunk_info(lat, lon, nc_data)
}
\arguments{
\item{lat}{numeric}

\item{lon}{numeric}

\item{nc_data}{Generated from the load_index_data}
}
\value{
Vector containing chunk id, y and x
}
\description{
Utilized in the purrr function. This function will return metadata for a latitude and longitude in the specific chunk.
}
\examples{
# Set coordinates of farm field
farm_coords <- c(43.854, -111.776)
# Download HRRR grid metadata to get chunk_id and x and y for coordinates
nc_data <- load_index_data()
# Get chunk coordinates for farm field
chunk_info <- get_chunk_info(farm_coords[1], farm_coords[2], nc_data)
chunk_id <- chunk_info[1]
in_chunk_row <- chunk_info[2]
in_chunk_col <- chunk_info[3]
}
