% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{get_chunks}
\alias{get_chunks}
\title{Title}
\usage{
get_chunks(x, max_date, min_date, rows = NULL)
}
\arguments{
\item{max_date}{A character}

\item{min_date}{A Character}

\item{rows}{defaults to null}
}
\description{
Dates are converted to Mountain Standard time zone. It gets a sequence of hours from the date range and combines into a dataframe.
If rows are not null, it Utilizes purrr's parallel processing when it calls read_grid_from_url. It then takes the results from read_grid_from_url
and converts to a tibble inside a dataframe.
}
