% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{get_previous_date}
\alias{get_previous_date}
\title{Gets Previous Days Date}
\usage{
get_previous_date()
}
\value{
a Date object
}
\description{
Gets Previous Days Date
}
\examples{
yesterday <- get_previous_date()
}
