% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{get_temp_from_date_range_cdat}
\alias{get_temp_from_date_range_cdat}
\title{Title}
\usage{
get_temp_from_date_range_cdat(fields, c_id, in_x, in_y, iteration)
}
\arguments{
\item{iteration}{}
}
\description{
Title
}
