% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpolation_test_dataset}
\alias{interpolation_test_dataset}
\title{Interpolation Test Dataset}
\format{
a dataframe of 200 rows and 8 columns
\describe{
\item{date}{Date}
\item{Lat}{Latitude}
\item{Long}{Longitude}
\item{GDD_alfalfa}{The GDD for Alfalfa}
\item{is_na}{Binary column with 1 respresenting na value in GDD calculation}
\item{na_ma_method}{Weighted Moving Average results}
\item{na_interpolation_method}{Linear Average results}
\item{GDD_alfalfa_answer}{Contains true value where Na is in GDD_alfalfa}
}
}
\usage{
interpolation_test_dataset
}
\description{
A small dataset derieved from the data retrieval process to test which method of interpolation to use. NA values were made on random values. The dataset
Contains the true values associated with the na values. The methods tested were Linear or Weighted Moving average.
}
\keyword{datasets}
