% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_calculations.R
\name{li}
\alias{li}
\title{Linear Interpolation of GDD}
\usage{
li(data)
}
\arguments{
\item{dataset}{}
}
\value{
dataset
}
\description{
li inputs is a dataframe that has the GDD over a set of days. It then groups by
Field ID, crop season, and cropname where each group is split into a dataframe. Linear interpolation
is done on each dataframe to replace those na values. It then rbinds each dataframe back together to return
a dataframe with linear interpolation applied and a new column that identifies those days with na values

It should be noted when an na value is on the last or first row of the dataframe the linear interpolation will
assign the na value the value of the next column. So the previous if the na value is the last row or second column
if it is the first row.
}
\examples{

# By Viewing you will see two na values
View(li_test)

li_data <- li(li_test)
# The Results of the data after Linear interpolation is applied
View(li_data)
}
