% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDH_calculations.R
\name{li_na_count}
\alias{li_na_count}
\title{Percent of days with NA for temperature}
\usage{
li_na_count(data)
}
\arguments{
\item{data}{}
}
\value{
Statement with the percent of days that contain na values for temperature
}
\description{
Percent of days with NA for temperature
}
\examples{

# View(li_test)

li_na_count(li_test)

}
