% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{li_test}
\alias{li_test}
\title{Linear Interpolation}
\format{
A data frame with 8 rows and 9 columns:
\describe{
\item{Field ID}{The field ID of interest}
\item{Crop Season}{Year for the crop}
\item{lat}{Latitude of the field}
\item{long}{Longitude of the field}
\item{seeding_date}{Date the seed was planted}
\item{harvest_date}{Date the seed will be harvested}
\item{date_02}{A date during the specified seeding and harvest date}
\item{GDD}{A number representing the GDD calculated for that day}
}
}
\source{
Data was generated and is found in the DATASET.R script
}
\usage{
li_test
}
\description{
An example dataset resembling the input from the csv file and output after the GDD function has been applied. It is
designed to show how the Linear interpolation works.
}
\keyword{datasets}
