% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{match_combine_data}
\alias{match_combine_data}
\title{Title}
\usage{
match_combine_data(fields_chunk_info, cdat)
}
\arguments{
\item{cdat}{}
}
\value{
a Dataframe containing temperatures by
}
\description{
takes fields chunk data and filters to seeding_date, harvest_date, chunk_id, in_chunk_x, in_chunk_y
then gets the temperature for the data by utilizing purrr's pmap function on the get_temp_from_date_range_cdat function.
}
