% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{purrf}
\alias{purrf}
\title{Utilizes parallel programming}
\usage{
purrf(la, lo)
}
\arguments{
\item{la}{Numeric}

\item{lo}{Numeric}
}
\value{
Vector containing chunk id, y and x
}
\description{
Calls the get_chunk_info and allows for parallel programming of the function so it gets all metadata of the specified chunk which
is a 150 by 150 grid.
}
