% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{read_grid_from_url}
\alias{read_grid_from_url}
\title{Reading Data from URL returned by get_url function}
\usage{
read_grid_from_url(hrrr_url)
}
\arguments{
\item{hrrr_url}{Character}
}
\value{
150 x 150 dataframe
}
\description{
If file doesn't exist it will set the temperature, in Kelvin (Temperature returned), to a na value. The
input of the function comes from the get_url function, but can get specified manually. See description for get_url to
know url.
}
\examples{
# Get a grid of temperature data for grid chunk 5.3 on June 24, 2022 at 3 PM MDT
hrrr_url <- get_url(
    as.POSIXct("2022-06-24 15:00:00", format="\%Y-\%m-\%d \%H:\%M:\%S", tz="US/Mountain"),
    "5.3")
# The URL is  "https://hrrrzarr.s3.amazonaws.com/sfc/20220624/20220624_15z_anl.zarr/surface/TMP/surface/TMP/5.3"
data_grid <- read_grid_from_url(hrrr_url)
}
