% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_data_hours}
\alias{temp_data_hours}
\title{Temperature Data Hours}
\format{
A dataframe of 1 row and 27 columns
\describe{
\item{crop_name}{Crop Name}
\item{lat}{Latitude}
\item{long}{Longitude}
\item{hour0, hour1, hour2, ...}{Hours of the day from hour 0 to hour 23}
}
}
\usage{
temp_data_hours
}
\description{
A dataset created for the GDD vignette that resembles the actual input data to the function in the script.
}
\keyword{datasets}
