% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{get_temp_from_date_cdat}
\alias{get_temp_from_date_cdat}
\title{Title}
\usage{
get_temp_from_date_cdat(gdh_date, chunk_id, in_chunk_x, in_chunk_y)
}
\arguments{
\item{in_chunk_y}{}
}
\description{
Title
}
