% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{load_index_data}
\alias{load_index_data}
\title{Gets Metadata on chunks from HRRR URL}
\usage{
load_index_data()
}
\value{
An object of class ncdf4. Format NC_FORMAT_64BIT
}
\description{
Data is pulled from https://aws.amazon.com/marketplace/pp/prodview-yd5ydptv3vuz2#resources. From this url,
"The HRRR is a NOAA real-time 3-km resolution, hourly updated, cloud-resolving, convection-allowing atmospheric model,
initialized by 3km grids with 3km radar assimilation. Radar data is assimilated in the HRRR every 15 min over a 1-h period
adding further detail to that provided by the hourly data assimilation from the 13km radar-enhanced Rapid Refresh."
}
\examples{
# Download HRRR grid metadata to get chunk_id and x and y for coordinates
nc_data <- load_index_data()
}
