% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrival.R
\name{purrf}
\alias{purrf}
\title{Title}
\usage{
purrf(la, lo)
}
\arguments{
\item{la}{Numeric}

\item{lo}{Numeric}
}
\value{
Vector containing chunk id, y and x
}
\description{
Calls the get_chunk_info and allows for parallization of the function. Has the ability to get all of the chunk_ids at one go.
}
